requirements.txt
numpy
opencv-python
torchaudio
torch
pydub
tqdm
scikit-learn
matplotlib
librosa
Pillow
transformers
pygame
speechbrain